Refactoring to suite complete project
