+++ api/src/main/resources/metadata/Locations.xml
@@ -2,5 +2,12 @@
 <dataset>
     <location location_id="9" name="Triage" creator="1" description="A location for categorization of patients" date_created="2019-09-16 12:04:24" retired="0"  parent_location="2"  uuid="ff01eaab-561e-40c6-bf24-539206b521ce"/>
     <location location_id="10" name="Reception"  description="A Location for registering patients" creator="1" date_created="2019-09-16 13:05:37" retired="0" parent_location="2"  uuid="4501e132-07a2-4201-9dc8-2f6769b6d412" />
-    <location  location_id="11"  name="Pharmacy"  description="A place for preparing, dispensing, and reviewing drugs and providing additional clinical services"  creator="1"  date_created="2019-09-16 13:13:31"  retired="0"  uuid="3ec8ff90-3ec1-408e-bf8c-22e4553d6e17"/>
+    <location  location_id="11"  name="Pharmacy"  description="A place for preparing, dispensing, and reviewing drugs and providing additional clinical services"  creator="1"  date_created="2019-09-16 13:13:31"  retired="0" parent_location="2"  uuid="3ec8ff90-3ec1-408e-bf8c-22e4553d6e17"/>
+
+    <location_tag_map  location_id="9"  location_tag_id="2"/>
+    <location_tag_map  location_id="9"  location_tag_id="4"/>
+    <location_tag_map  location_id="10"  location_tag_id="2"/>
+    <location_tag_map  location_id="10"  location_tag_id="4"/>
+    <location_tag_map  location_id="11"  location_tag_id="2"/>
+    <location_tag_map  location_id="11"  location_tag_id="4"/>
 </dataset>
