+++ api/src/main/resources/moduleApplicationContext.xml
@@ -24,4 +24,29 @@
   		    http://www.springframework.org/schema/util/spring-util-3.0.xsd">
 
     <context:component-scan base-package="org.openmrs.module.ugandaemrpoc" />
+    <bean id="ugandaemrpoc.UgandaEMRPOCService"
+          class="org.springframework.transaction.interceptor.TransactionProxyFactoryBean">
+        <property name="transactionManager">
+            <ref bean="transactionManager"/>
+        </property>
+        <property name="target">
+            <bean class="org.openmrs.module.ugandaemrpoc.api.impl.UgandaEMRPOCServiceImpl">
+            </bean>
+        </property>
+        <property name="preInterceptors">
+            <ref bean="serviceInterceptors"/>
+        </property>
+        <property name="transactionAttributeSource">
+            <ref bean="transactionAttributeSource"/>
+        </property>
+    </bean>
+
+    <bean parent="serviceContext">
+        <property name="moduleService">
+            <list>
+                <value>org.openmrs.module.ugandaemrpoc.api.UgandaEMRPOCService</value>
+                <ref bean="ugandaemrpoc.UgandaEMRPOCService"/>
+            </list>
+        </property>
+    </bean>
 </beans>
